requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml==5.1.0
python-telegram-bot==20.7
sqlalchemy>=2.0.0
pandas>=2.0.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.21.0
pyarrow>=23.0.0
aiohttp==3.9.3
asyncio==3.4.3
